feat: Enable thread submission using the Enter key
This change updates the key handling logic within the `Update` function to enable thread submission consistently using the "Enter" key. Previously, submitting a thread with the "Enter" key was restricted to when the input focus was specifically on `FocusedInput`. With this update, pressing "Enter" will submit the current message and save the thread regardless of the focus.

Activating a thread is now done using the space key.